move example to sample.py, modify app.py to be more relevant
